Moved EmptyPlayerStateTests to separated class
